+++ packages/vue/src/components/menu/menu-checkbox-item.vue
@@ -42,9 +42,9 @@ const optionItemProps: ComputedRef<OptionItemProps> = computed(() => ({
   onCheckedChange: (e) => emits('update:checked', e),
 }))
 
-const item = computed(() => menu.value.getOptionItemState(optionItemProps.value))
+const optionItemState = computed(() => menu.value.getOptionItemState(optionItemProps.value))
 
-MenuItemProvider(item)
+MenuItemProvider(optionItemState)
 MenuOptionItemPropsProvider(optionItemProps)
 </script>