+++ packages/solid/CHANGELOG.md
@@ -23,6 +23,7 @@ export interface AccordionItemProps extends HTMLProps<'div'>, AccordionItemBaseP
 - **Menu:** 
   - Resolved issue causing sibling menus or popovers to malfunction on iOS devices.
   - Fixed problem where the context menu trigger showed the magnifier and iOS context menu on long press.
+  - Fixed issue where `RadioOptionItem` does not provide the `checked` state in context.
 - **Editable:** Fixed bug where setting `activationMode` to `dblclick` unexpectedly cleared the input value.
 - **Checkbox:** Added the missing `group` role for the Checkbox group component.
 - **Signature Pad**: Added missing package.json for `signature-pad` machine.